Move/rename spatial.find_geom_in_swn to locate_geoms instance method
