+++ docs/source/spatial.rst
@@ -2,10 +2,13 @@ Spatial methods
 ===============
 
 .. module:: swn.spatial
+
+Paired data methods
+-------------------
+
 .. autosummary::
    :toctree: ref/
 
-   find_geom_in_swn
    find_location_pairs
    location_pair_geoms
